Added required dataset,images,debts crud,fixing some issues  and many more

+++ Helper/deleteAfterMulter.js
@@ -2,10 +2,12 @@
 const fs = require('fs');
 
 module.exports = (imagePath) => {
-    fs.unlink(imagePath, (err)=>{
-        if(err)
-            console.log(err)
-        else
-            console.log('deleted');
-    });
+    if(fs.existsSync(imagePath)){
+        fs.unlink(imagePath, (err)=>{
+            if(err)
+                console.log(err)
+            else
+                console.log('deleted');
+        });
+    }
 };
